third_party/blink/web_tests/http/tests/security/base-url-javascript-invalid.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <base href="javascript:// This is an invalid JavaScript URL">
</head>
<body>
  <script>
    async_test(t => {
      const img = new Image();
      img.onload = t.unreached_func("Image should not have been loaded.");
      img.onerror =  t.step_func_done(() => {});
      img.src = "/security/resources/abe.png";
    }, "A base URL is an invalid 'javascript:' URL, whose host part includes spaces.");
  </script>
</body>
